requirements.txt
# Core ML and data processing
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
numpy>=1.24.0
joblib>=1.3.0
lightgbm>=4.0.0
catboost>=1.2.0

# API and utilities
requests>=2.31.0
python-dotenv>=1.0.0

# Supabase client (lightweight alternative)
postgrest>=0.10.0
httpx>=0.24.0

# UI
streamlit>=1.25.0
